documenting
git-svn-id: svn://localhost/trunk@709 3bdad9f7-c964-442a-868d-553b97b55ca0

+++ surfit/src/sstuff/fileio.cpp
@@ -261,7 +261,7 @@ int calc_columns(const char * Str, int readed) {
 
 	char * token = strtok( str, seps );
 	while (token != NULL) {
-		REAL val = atof(token);
+		REAL val = (REAL)atof(token);
 		if (val != 0)
 			columns++;
 		else {
